% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_nlopt_c}
\alias{test_nlopt_c}
\title{NLOPT Integration Example}
\usage{
test_nlopt_c()
}
\value{
A numeric vector with two elements
}
\description{
A simple example for for NLOPT integration for RcppArmadillo
}
\examples{
test_nlopt_c()
}
