% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{testConstrainedProblem}
\alias{testConstrainedProblem}
\title{NLopt Call Example from Rcpp}
\usage{
testConstrainedProblem(method, verbose = FALSE)
}
\arguments{
\item{method}{A character value selecting the optimization methods}

\item{verbose}{A boolean toggle defaulting to \sQuote{false}}
}
\value{
A numeric vector with two elements
}
\description{
A simple example for for NLopt integration for Rcpp,
using an example from the NLopt tutorial.
}
\examples{
testConstrainedProblem("MMA", TRUE)
}
\seealso{
\url{https://nlopt.readthedocs.io/en/latest/NLopt_Tutorial/}
}
